// src/notification/enums/comms-event-type.enum.ts
export enum CommsEventType {
  ORDER_CONFIRM = 'ORDER_CONFIRM',
  ORDER_SHIPPED = 'ORDER_SHIPPED',
  ORDER_DELIVERED = 'ORDER_DELIVERED',
  ORDER_CANCELLED = 'ORDER_CANCELLED',
  ORDER_FAILED = 'ORDER_FAILED',
  RETURN_INITIATED = 'RETURN_INITIATED',
  RETURN_CANCELLED = 'RETURN_CANCELLED',
  EXCHANGE_INITIATED = 'EXCHANGE_INITIATED',
  EXCHANGE_CANCELLED = 'EXCHANGE_CANCELLED',
  REFUND_INITIATED = 'REFUND_INITIATED',
}